Modularize project structure: split functionality across multiple files and headers; update CMakeLists.txt accordingly.
