+++ src/include/csrf.h
@@ -1,0 +1,8 @@
+//
+// Created by olegek on 6/13/25.
+//
+
+#ifndef CSRF_H
+#define CSRF_H
+
+#endif //CSRF_H
