+++ src/include/menu.h
@@ -1,0 +1,17 @@
+//
+// Created by olegek on 6/13/25.
+//
+
+#ifndef BANNER_H
+#define BANNER_H
+
+// Печатает баннер с ASCII-арт и информацией о проекте
+void print_banner();
+
+// Печатает главное меню с баннером и опциями
+void print_start_menu();
+
+// Очищает вывод консоли
+void clear_console();
+
+#endif //BANNER_H
