+++ src/include/pch.h
@@ -1,0 +1,8 @@
+#ifndef PCH_H
+#define PCH_H
+
+#include <iostream>
+#include <string>
+#include <vector>
+
+#endif //PCH_H
