+++ src/include/sqlinj.h
@@ -1,0 +1,8 @@
+//
+// Created by olegek on 6/13/25.
+//
+
+#ifndef SQLINJ_H
+#define SQLINJ_H
+
+#endif //SQLINJ_H
