+++ src/include/xss.h
@@ -1,0 +1,8 @@
+//
+// Created by olegek on 6/13/25.
+//
+
+#ifndef XSS_H
+#define XSS_H
+
+#endif //XSS_H
