+++ src/menu.cpp
@@ -1,0 +1,66 @@
+
+#include "include/menu.h"
+
+// Очищает экран консоли (с помощью полной перезагрузки терминала)
+void clear_console() {
+    std::cout << "\033c" << std::flush;
+}
+
+// Печатает баннер с ASCII-арт и информацией о проекте
+void print_banner() {
+    static const char* banner = R"(
+     @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
+  @n                                                           @@
+@@                                                               @
+@                                        @                        @
+@     @@@@@@       @@@@              @@@@@@                         @
+@      @@@@@       @@@                 @@@@                         @
+@       @@@@@     @@@                  @@@@                         @
+@       @@@@@     @@@ nnnnn   nnnnn    @@@@     @@@  @@@@           @
+@        @@@@n   n@@   @@@@n   @@@@    @@@@   @@@@@@@@@@@@          @
+@        @@@@@   @@@   @@@@n   @@@@    @@@@   t@@@@   @@@@@         @
+@         @@@@   @@    @@@@n   @@@@    @@@@    @@@@    @@@@         @
+@         @@@@@ @@@    @@@@n   @@@@    @@@@    @@@@    @@@@         @
+@          @@@@ @@     @@@@n   @@@@    @@@@    @@@@    @@@@         @
+@          @@@@@@@     @@@@6   @@@@    @@@@    @@@@    @@@@         @
+@           @@@@@      @@@@@   @@@@    @@@@    @@@@    @@@@         @
+@           @@@@@       @@@@@@@n@@@@   @@@@   @@@@@   @@@@@         @
+@                         n@n                                       @
+@            @@@@   @                                               @
+@         @@@@@@@@@@@                                               @
+@        @@@@      @@                                               @
+@        @@@@       @        @          nn           @   6@         @
+@        @@@@@           @@@@@@@@@   @@@@@@@@    @@@@@n@@@@@@       @
+@        @@@@@@@@       @@@@    @@  @@    @@@@    @@@@n  @@@@@      @
+@          @@@@@@@@n   @@@@      @        @@@@    @@@@    @@@@      @
+@            n@@@@@@@  @@@@             @@@@@@    @@@@    @@@@      @
+@               @@@@@@ @@@@         @@@@  @@@@    @@@@    @@@@      @
+@        @       @@@@@ @@@@@       @@@@   @@@@    @@@@    @@@@      @
+@        @@      @@@@@  @@@@      n@@@@   @@@@    @@@@    @@@@      @
+@        @@@@n  @@@@@   @@@@@@@@@@ @@@@@@@@@@@    @@@@    @@@@      @
+@        @@@@@@@@@@       @@@@@@    @@@@@ n@@@@@ @@@@@@  @@@@@@     @
+@                                                                   @
+ @                                                                 @
+  n@                                                           @
+       @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@t
+
+
+   [---]        The Vulnerability Scanner (VulnScan)         [---]
+[---]        Created by: Usoltsev, Yalfimov and Korolev         [---]
+                           Version: 1.0.0
+     [---] Visit: https://github.com/maxim290905/VulnScan [---]
+                    [---] Telegram: @pvppve [---]
+
+)";
+    std::cout << banner;
+}
+
+// Печатает главное меню с баннером и опциями
+void print_start_menu() {
+    print_banner();
+    static const char* inform = R"(
+Welcome to VulnScan!
+Enter URL:
+)";
+    std::cout << inform;
+}
